14/4 18h20 thêm, sửa, xóa, tìm kiếm (theo khoảng ngày, nguyên liệu, nhà cung cấp, tổng hợp) phiếu nhập kho; báo cáo tồn kho; cảnh báo nguyên liệu sắp hết; bỏ xử lí nhà cung cấp

+++ backend/src/main/java/com/fastfood/dto/response/InventoryItemReportResponse.java
@@ -1,6 +1,12 @@
 package com.fastfood.dto.response;
 
-import lombok.*;
+import java.math.BigDecimal;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 import lombok.experimental.FieldDefaults;
 
 @Data
@@ -12,7 +18,7 @@ public class InventoryItemReportResponse {
     String ingredientId;
     String ingredientName;
     String unit;
-    Integer currentStock;
-    Integer minStock;
+    BigDecimal currentStock;
+    BigDecimal minStock;
     boolean lowStock;
 }
